update cluster autoscaling configs
add new clusterrole entries policy rules

set cluster autoscaler image to kubernetes version
